refactor: create hook for logic
